Fix filename refactor
