+++ ingestion_tools/scripts/data_validation/standardized/tests/test_alignments.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from data_validation.standardized.tests.helper_angles import helper_angles_injection_errors
+from data_validation.shared.helper.angles_helper import helper_angles_injection_errors
 
 
 @pytest.mark.alignment
