+++ ingestion_tools/scripts/data_validation/standardized/tests/test_frame.py
@@ -8,7 +8,7 @@ import pytest
 import tifffile
 from mrcfile.mrcinterpreter import MrcInterpreter
 
-from data_validation.shared.helper.frame_helper import FrameTestHelper
+from data_validation.shared.helper.twodee_helper import FrameTestHelper
 
 
 @pytest.mark.frame
